chore(evm): reorganize `foundry-evm-core::evm` mod
